+++ forgerock-auth/src/main/java/org/forgerock/android/auth/FRUser.java
@@ -8,9 +8,13 @@
 package org.forgerock.android.auth;
 
 import android.content.Context;
+
 import androidx.annotation.WorkerThread;
-import lombok.RequiredArgsConstructor;
+
+import org.forgerock.android.auth.exception.AlreadyAuthenticatedException;
 import org.forgerock.android.auth.exception.AuthenticationRequiredException;
+
+import lombok.RequiredArgsConstructor;
 
 public class FRUser {
 
@@ -97,35 +101,56 @@ public class FRUser {
     }
 
     /**
-     * Trigger the user login process, the login service name can be defined under <b>string.xml</b> file with
+     * Trigger the user login process, the login service name is defined under <b>string.xml</b> file with
      * <b>forgerock_auth_service</b>
      *
      * @param context  The Application Context
      * @param listener Listener to listen login event.
+     * <b> Throw {@link AlreadyAuthenticatedException} user session already exists.
      */
     public static void login(Context context, final NodeListener<FRUser> listener) {
-        createFRAuth(context, context.getString(R.string.forgerock_auth_service))
-                .next(context, listener);
-    }
+        SessionManager sessionManager = SessionManager.builder()
+                .build();
 
-    private static FRAuth createFRAuth(Context context, String serviceName) {
+        if (sessionManager.hasSession() ) {
+            Listener.onException(listener, new AlreadyAuthenticatedException("User is already authenticated"));
+            return;
+        }
+
+        createFRAuth(context, context.getString(R.string.forgerock_auth_service), sessionManager)
+                .next(context, listener);
+
+   }
+
+    private static FRAuth createFRAuth(Context context, String serviceName, SessionManager sessionManager) {
         return FRAuth.builder()
                 .serviceName(serviceName)
                 .context(context)
+                .interceptor(new OAuthInterceptor(sessionManager.getOAuth2Client()))
+                .interceptor(new AccessTokenStoreInterceptor(sessionManager.getTokenManager()))
                 .interceptor(new UserInterceptor())
                 .build();
     }
 
     /**
-     * Trigger the user login process, the registration service name can be defined under <b>string.xml</b> file with
+     * Trigger the user registration process, the registration service name is defined under <b>string.xml</b> file with
      * <b>forgerock_registration_service</b>
      *
      * @param context  The Application Context
-     * @param listener Listener to listen login event.
+     * @param listener Listener to listen register event.
+     * <b> Throw {@link AlreadyAuthenticatedException} user session already exists.
      */
     public static void register(Context context, NodeListener<FRUser> listener) {
-        createFRAuth(context, context.getString(R.string.forgerock_registration_service))
-                .start(context, listener);
+        SessionManager sessionManager = SessionManager.builder()
+                .build();
+
+        if (sessionManager.hasSession() ) {
+            Listener.onException(listener, new AlreadyAuthenticatedException("User is already authenticated"));
+            return;
+        }
+
+        createFRAuth(context, context.getString(R.string.forgerock_registration_service), sessionManager)
+                .next(context, listener);
     }
 
     @RequiredArgsConstructor