+++ forgerock-auth/src/test/java/org/forgerock/android/auth/AuthTestSuite.java
@@ -38,6 +38,7 @@ import org.junit.runners.Suite;
         DefaultSingleSignOnManagerTest.class,
 
         FRUserMockTest.class,
+        FRSessionMockTest.class,
         RootDetectorTest.class,
         FRDeviceTest.class,