+++ forgerock-auth/src/test/java/org/forgerock/android/auth/FRAuthMockTest.java
@@ -42,10 +42,6 @@ public class FRAuthMockTest extends BaseTest {
         enqueue("/authTreeMockTest_Authenticate_NameCallback.json", HttpURLConnection.HTTP_OK);
         enqueue("/authTreeMockTest_Authenticate_PasswordCallback.json", HttpURLConnection.HTTP_OK);
         enqueue("/authTreeMockTest_Authenticate_success.json", HttpURLConnection.HTTP_OK);
-        server.enqueue(new MockResponse()
-                .addHeader("Location", "http://www.example.com:8080/callback?code=PmxwECH3mBobKuPEtPmq6Xorgzo&iss=http://openam.example.com:8080/openam/oauth2&state=abc123&client_id=andy_app")
-                .setResponseCode(HttpURLConnection.HTTP_MOVED_TEMP));
-        enqueue("/authTreeMockTest_Authenticate_accessToken.json", HttpURLConnection.HTTP_OK);
 
         final SingleSignOnManager singleSignOnManager = DefaultSingleSignOnManager.builder()
                 .context(context)
@@ -56,8 +52,6 @@ public class FRAuthMockTest extends BaseTest {
                 .serviceName("Example")
                 .context(context)
                 .sessionManager(SessionManager.builder()
-                        .oAuth2Client(oAuth2Client)
-                        .tokenManager(new DoNothingTokenManager())
                         .singleSignOnManager(singleSignOnManager)
                         .build())
                 .serverConfig(serverConfig)
@@ -82,7 +76,7 @@ public class FRAuthMockTest extends BaseTest {
 
         frAuth.next(context, nodeListenerFuture);
 
-        Assert.assertTrue(nodeListenerFuture.get() instanceof AccessToken);
+        Assert.assertTrue(nodeListenerFuture.get() instanceof SSOToken);
 
     }
 
@@ -92,10 +86,6 @@ public class FRAuthMockTest extends BaseTest {
         enqueue("/authTreeMockTest_Authenticate_NameCallback.json", HttpURLConnection.HTTP_OK);
         enqueue("/authTreeMockTest_Authenticate_PasswordCallback.json", HttpURLConnection.HTTP_OK);
         enqueue("/authTreeMockTest_Authenticate_success.json", HttpURLConnection.HTTP_OK);
-        server.enqueue(new MockResponse()
-                .addHeader("Location", "http://www.example.com:8080/callback?code=PmxwECH3mBobKuPEtPmq6Xorgzo&iss=http://openam.example.com:8080/openam/oauth2&state=abc123&client_id=andy_app")
-                .setResponseCode(HttpURLConnection.HTTP_MOVED_TEMP));
-        enqueue("/authTreeMockTest_Authenticate_accessToken.json", HttpURLConnection.HTTP_OK);
 
         // Mocking the config
         Resources resources = Mockito.mock(Resources.class);
@@ -121,8 +111,6 @@ public class FRAuthMockTest extends BaseTest {
                 .serviceName("Example")
                 .context(mockContext)
                 .sessionManager(SessionManager.builder()
-                        .oAuth2Client(oAuth2Client)
-                        .tokenManager(new DoNothingTokenManager())
                         .singleSignOnManager(singleSignOnManager)
                         .build())
                 .serverConfig(serverConfig)
@@ -147,7 +135,7 @@ public class FRAuthMockTest extends BaseTest {
 
         frAuth.next(mockContext, nodeListenerFuture);
 
-        Assert.assertTrue(nodeListenerFuture.get() instanceof AccessToken);
+        Assert.assertTrue(nodeListenerFuture.get() instanceof SSOToken);
 
         RecordedRequest rr = server.takeRequest(); //Start the Auth Service
         assertEquals("/json/realms/root/authenticate?authIndexType=service&authIndexValue=Example", rr.getPath());
@@ -158,30 +146,14 @@ public class FRAuthMockTest extends BaseTest {
         rr = server.takeRequest(); //Post Password Callback
         assertEquals("/json/realms/root/authenticate", rr.getPath());
 
-        rr = server.takeRequest(); //Post to /authorize endpoint
-        assertTrue(rr.getPath().startsWith("/oauth2/realms/root/authorize?iPlanetDirectoryPro="));
-
-        rr = server.takeRequest(); //Post to /access-token endpoint
-        assertEquals("/oauth2/realms/root/access_token", rr.getPath());
-
     }
 
     @Test
-    public void testAccessTokenWithSharedPreferencesSSOManager() throws ExecutionException, InterruptedException, AuthenticationRequiredException {
+    public void testWithSharedPreferencesSSOManager() throws ExecutionException, InterruptedException, AuthenticationRequiredException {
 
         enqueue("/authTreeMockTest_Authenticate_NameCallback.json", HttpURLConnection.HTTP_OK);
         enqueue("/authTreeMockTest_Authenticate_PasswordCallback.json", HttpURLConnection.HTTP_OK);
         enqueue("/authTreeMockTest_Authenticate_success.json", HttpURLConnection.HTTP_OK);
-        server.enqueue(new MockResponse()
-                .addHeader("Location", "http://www.example.com:8080/callback?code=PmxwECH3mBobKuPEtPmq6Xorgzo&iss=http://openam.example.com:8080/openam/oauth2&state=abc123&client_id=andy_app")
-                .setResponseCode(HttpURLConnection.HTTP_MOVED_TEMP));
-        enqueue("/authTreeMockTest_Authenticate_accessToken.json", HttpURLConnection.HTTP_OK);
-
-        final TokenManager tokenManager = DefaultTokenManager.builder()
-                .context(context)
-                .oAuth2Client(oAuth2Client)
-                .sharedPreferences(context.getSharedPreferences("Test", Context.MODE_PRIVATE))
-                .build();
 
         final SingleSignOnManager singleSignOnManager = SharedPreferencesSignOnManager.builder()
                 .context(context)
@@ -192,8 +164,6 @@ public class FRAuthMockTest extends BaseTest {
                 .serviceName("Example")
                 .context(context)
                 .sessionManager(SessionManager.builder()
-                        .oAuth2Client(oAuth2Client)
-                        .tokenManager(tokenManager)
                         .singleSignOnManager(singleSignOnManager)
                         .build())
                 .serverConfig(serverConfig)
@@ -218,254 +188,11 @@ public class FRAuthMockTest extends BaseTest {
 
         frAuth.next(context, nodeListenerFuture);
 
-        Assert.assertTrue(nodeListenerFuture.get() instanceof AccessToken);
-
-        //Check AccessToken Storage
-        AccessToken accessToken = tokenManager.getAccessToken();
-        assertNotNull(accessToken);
-        assertNotNull(accessToken.getValue());
+        Assert.assertTrue(nodeListenerFuture.get() instanceof SSOToken);
 
         //Check SSOToken Storage
         Token token = singleSignOnManager.getToken();
         assertNotNull(token);
-    }
-
-    @Test
-    public void testAccessTokenAndSSOToken() throws ExecutionException, InterruptedException, AuthenticationRequiredException {
-
-        enqueue("/authTreeMockTest_Authenticate_NameCallback.json", HttpURLConnection.HTTP_OK);
-        enqueue("/authTreeMockTest_Authenticate_PasswordCallback.json", HttpURLConnection.HTTP_OK);
-        enqueue("/authTreeMockTest_Authenticate_success.json", HttpURLConnection.HTTP_OK);
-        server.enqueue(new MockResponse()
-                .addHeader("Location", "http://www.example.com:8080/callback?code=PmxwECH3mBobKuPEtPmq6Xorgzo&iss=http://openam.example.com:8080/openam/oauth2&state=abc123&client_id=andy_app")
-                .setResponseCode(HttpURLConnection.HTTP_MOVED_TEMP));
-        enqueue("/authTreeMockTest_Authenticate_accessToken.json", HttpURLConnection.HTTP_OK);
-
-        final TokenManager tokenManager = DefaultTokenManager.builder()
-                .context(context)
-                .oAuth2Client(oAuth2Client)
-                .sharedPreferences(context.getSharedPreferences("Test", Context.MODE_PRIVATE))
-                .build();
-
-        final SingleSignOnManager singleSignOnManager = DefaultSingleSignOnManager.builder()
-                .context(context)
-                .encryptor(new MockEncryptor())
-                .build();
-
-        final FRAuth frAuth = FRAuth.builder()
-                .serviceName("Example")
-                .context(context)
-                .sessionManager(SessionManager.builder()
-                        .oAuth2Client(oAuth2Client)
-                        .tokenManager(tokenManager)
-                        .singleSignOnManager(singleSignOnManager)
-                        .build())
-                .serverConfig(serverConfig)
-                .build();
-
-        NodeListenerFuture nodeListenerFuture = new NodeListenerFuture() {
-
-            @Override
-            public void onCallbackReceived(Node state) {
-                if (state.getCallback(NameCallback.class) != null) {
-                    state.getCallback(NameCallback.class).setName("tester");
-                    state.next(context, this);
-                    return;
-                }
-
-                if (state.getCallback(PasswordCallback.class) != null) {
-                    state.getCallback(PasswordCallback.class).setPassword("password".toCharArray());
-                    state.next(context, this);
-                }
-            }
-        };
-
-        frAuth.next(context, nodeListenerFuture);
-
-        Assert.assertTrue(nodeListenerFuture.get() instanceof AccessToken);
-
-        //Check AccessToken Storage
-        AccessToken accessToken = tokenManager.getAccessToken();
-        assertNotNull(accessToken);
-        assertNotNull(accessToken.getValue());
-
-        //Check SSOToken Storage
-        Token token = singleSignOnManager.getToken();
-        assertNotNull(token);
-        assertNotNull(token.getValue());
-    }
-
-    @Test
-    public void testAccessTokenAndSSOTokenRefreshWithSSOToken() throws ExecutionException, InterruptedException, AuthenticationRequiredException {
-
-        enqueue("/authTreeMockTest_Authenticate_NameCallback.json", HttpURLConnection.HTTP_OK);
-        enqueue("/authTreeMockTest_Authenticate_PasswordCallback.json", HttpURLConnection.HTTP_OK);
-        enqueue("/authTreeMockTest_Authenticate_success.json", HttpURLConnection.HTTP_OK);
-        server.enqueue(new MockResponse()
-                .addHeader("Location", "http://www.example.com:8080/callback?code=PmxwECH3mBobKuPEtPmq6Xorgzo&iss=http://openam.example.com:8080/openam/oauth2&state=abc123&client_id=andy_app")
-                .setResponseCode(HttpURLConnection.HTTP_MOVED_TEMP));
-        enqueue("/authTreeMockTest_Authenticate_accessToken.json", HttpURLConnection.HTTP_OK);
-        server.enqueue(new MockResponse()
-                .addHeader("Location", "http://www.example.com:8080/callback?code=PmxwECH3mBobKuPEtPmq6Xorgzo&iss=http://openam.example.com:8080/openam/oauth2&state=abc123&client_id=andy_app")
-                .setResponseCode(HttpURLConnection.HTTP_MOVED_TEMP));
-        enqueue("/authTreeMockTest_Authenticate_accessToken.json", HttpURLConnection.HTTP_OK);
-
-        final TokenManager tokenManager = DefaultTokenManager.builder()
-                .context(context)
-                .oAuth2Client(oAuth2Client)
-                .sharedPreferences(context.getSharedPreferences("Test", Context.MODE_PRIVATE))
-                .build();
-
-        final SingleSignOnManager singleSignOnManager = DefaultSingleSignOnManager.builder()
-                .context(context)
-                .encryptor(new MockEncryptor())
-                .build();
-
-        SessionManager sessionManager = SessionManager.builder()
-                .tokenManager(tokenManager)
-                .singleSignOnManager(singleSignOnManager)
-                .oAuth2Client(oAuth2Client)
-                .build();
-
-
-        final FRAuth frAuth = FRAuth.builder()
-                .serviceName("Example")
-                .context(context)
-                .sessionManager(sessionManager)
-                .serverConfig(serverConfig)
-                .build();
-
-        NodeListenerFuture nodeListenerFuture = new NodeListenerFuture() {
-
-            @Override
-            public void onCallbackReceived(Node state) {
-                if (state.getCallback(NameCallback.class) != null) {
-                    state.getCallback(NameCallback.class).setName("tester");
-                    state.next(context, this);
-                    return;
-                }
-
-                if (state.getCallback(PasswordCallback.class) != null) {
-                    state.getCallback(PasswordCallback.class).setPassword("password".toCharArray());
-                    state.next(context, this);
-                }
-            }
-        };
-
-        frAuth.next(context, nodeListenerFuture);
-
-        Assert.assertTrue(nodeListenerFuture.get() instanceof AccessToken);
-
-        //Check AccessToken Storage
-        AccessToken accessToken = tokenManager.getAccessToken();
-        assertNotNull(accessToken);
-        assertNotNull(accessToken.getValue());
-
-        //Check SSOToken Storage
-        Token token = singleSignOnManager.getToken();
-        assertNotNull(token);
-        assertNotNull(token.getValue());
-
-        tokenManager.clear();
-
-        accessToken = sessionManager.getAccessToken();
-        assertNotNull(accessToken.getValue());
-
-
-    }
-
-    @Test
-    public void testSSOEnabled() throws ExecutionException, InterruptedException, AuthenticationRequiredException {
-
-        enqueue("/authTreeMockTest_Authenticate_NameCallback.json", HttpURLConnection.HTTP_OK);
-        enqueue("/authTreeMockTest_Authenticate_PasswordCallback.json", HttpURLConnection.HTTP_OK);
-        enqueue("/authTreeMockTest_Authenticate_success.json", HttpURLConnection.HTTP_OK);
-        server.enqueue(new MockResponse()
-                .addHeader("Location", "http://www.example.com:8080/callback?code=PmxwECH3mBobKuPEtPmq6Xorgzo&iss=http://openam.example.com:8080/openam/oauth2&state=abc123&client_id=andy_app")
-                .setResponseCode(HttpURLConnection.HTTP_MOVED_TEMP));
-        enqueue("/authTreeMockTest_Authenticate_accessToken.json", HttpURLConnection.HTTP_OK);
-
-        //For AppB
-        server.enqueue(new MockResponse()
-                .addHeader("Location", "http://www.example.com:8080/callback?code=PmxwECH3mBobKuPEtPmq6Xorgzo&iss=http://openam.example.com:8080/openam/oauth2&state=abc123&client_id=andy_app")
-                .setResponseCode(HttpURLConnection.HTTP_MOVED_TEMP));
-        enqueue("/authTreeMockTest_Authenticate_accessToken.json", HttpURLConnection.HTTP_OK);
-
-        final TokenManager tokenManager = DefaultTokenManager.builder()
-                .context(context)
-                .oAuth2Client(oAuth2Client)
-                .sharedPreferences(context.getSharedPreferences("Test", Context.MODE_PRIVATE))
-                .build();
-
-        final SingleSignOnManager singleSignOnManager = DefaultSingleSignOnManager.builder()
-                .context(context)
-                .encryptor(new MockEncryptor())
-                .build();
-
-        final FRAuth frAuth = FRAuth.builder()
-                .serviceName("Example")
-                .context(context)
-                .sessionManager(SessionManager.builder()
-                        .oAuth2Client(oAuth2Client)
-                        .tokenManager(tokenManager)
-                        .singleSignOnManager(singleSignOnManager)
-                        .build())
-                .serverConfig(serverConfig)
-                .build();
-
-        NodeListenerFuture nodeListenerFuture = new NodeListenerFuture() {
-
-            @Override
-            public void onCallbackReceived(Node state) {
-                if (state.getCallback(NameCallback.class) != null) {
-                    state.getCallback(NameCallback.class).setName("tester");
-                    state.next(context, this);
-                    return;
-                }
-
-                if (state.getCallback(PasswordCallback.class) != null) {
-                    state.getCallback(PasswordCallback.class).setPassword("password".toCharArray());
-                    state.next(context, this);
-                }
-            }
-        };
-
-        frAuth.next(context, nodeListenerFuture);
-
-        Assert.assertTrue(nodeListenerFuture.get() instanceof AccessToken);
-
-        RecordedRequest rr = server.takeRequest(); //Start the Auth Service
-        rr = server.takeRequest(); //Post Name Callback
-        rr = server.takeRequest(); //Post Password Callback
-        rr = server.takeRequest(); //Post to /authorize endpoint
-        rr = server.takeRequest(); //Post to /access-token endpoint
-
-        //Switch to another App with Access Token does not exists
-        tokenManager.clear();
-
-        final boolean[] callbackReceived = new boolean[1];
-        NodeListenerFuture appB = new NodeListenerFuture() {
-            @Override
-            public void onCallbackReceived(Node node) {
-                callbackReceived[0] = true;
-            }
-        };
-        frAuth.next(context, appB);
-        assertFalse(callbackReceived[0]); //Should not receive callback
-        Assert.assertTrue(appB.get() instanceof AccessToken);
-        rr = server.takeRequest(); //Post to /authorize endpoint
-        rr = server.takeRequest(); //Post to /access-token endpoint
-
-        //Check AccessToken Storage
-        AccessToken accessToken = tokenManager.getAccessToken();
-        assertNotNull(accessToken);
-        assertNotNull(accessToken.getValue());
-
-        //Check SSOToken Storage
-        Token token = singleSignOnManager.getToken();
-        assertNotNull(token);
-        assertNotNull(token.getValue());
-
     }
 
     @Test